Revert "convert int to double variable.  display doble value with 2 decimal values"
This reverts commit 89984994a0fd76650c29a7e30f5f6de1b150d6d9.

+++ lib/widgets/add_transaction_form.dart
@@ -37,7 +37,7 @@ class _AddTransactionFormState extends State<AddTransactionForm> {
       });
       final user = FirebaseAuth.instance.currentUser;
       //int timestamp = DateTime.now().microsecondsSinceEpoch;
-      var amount = double.parse(amountEditController.text);
+      var amount = int.parse(amountEditController.text);
       DateTime date = DateTime.now();
 
       var id = uid.v4();
@@ -48,9 +48,9 @@ class _AddTransactionFormState extends State<AddTransactionForm> {
           .doc(user!.uid)
           .get();
 
-      double remainingAmount = userDoc['remainingAmount'].toDouble();
-      double totalCredit = userDoc['totalCredit'].toDouble();
-      double totalDebit = userDoc['totalDebit'].toDouble();
+      int remainingAmount = userDoc['remainingAmount'];
+      int totalCredit = userDoc['totalCredit'];
+      int totalDebit = userDoc['totalDebit'];
 
       if (type == 'Credit') {
         remainingAmount += amount;
